Revamped Ids.
We're going to have to keep project-external ids in name-based form because
they could resolve to different things depending on a given project's compile
time classpath.

Project A may use library B and dependency C1, and when A resolves a C from B
it wants to see C1's version of it. Project D may use library B and dependency
C2, (maybe a newer version of C1), so when it sees C via B it should see C2.
But there's no way for B to know which C we'll be talking about until it's
"instantiated" as part of project's A or D.