More id rethinking: Id -> Ref.
A Def always has a project local id because you always get a def from the
project in which it is defined. But a Use or a Relation may involve a non-local
def, so we provide Ref to indicate the referent of a use or relation. These may
be local or non-local.