Handle VarSymbol in owner chain of type.
Foo foo = new Foo() {
  ...
}

will result in 'foo' being the owner of the ClassSymbol for 'new Foo', and it
makes sense to nest 'new Foo' under the var to which it was assigned. I suppose
if a 'new Foo()' occurs inside a method without an lhs, we can just nest 'new
Foo' directly under the method.