Allow zip entries to be filtered.
Now I've decided it's better to just prune the JDK down to the 3000ish actual
java source files rather than compile the 7000ish files and only extract the
java/javax files.

It results in more warnings/errors, but there are already a bunch of those
because the JDK doesn't come with all of the s3cr3t stuffs needed to compile
itself anyway.