Added ProjectStore.usesOf(Def).
We currently maintain for each refs (local and global) that occurs anywhere in
a project, a listing of the compunits that contain that ref. When all
occurances of a ref are requested, we load that index of compunits and then
scan each compunit manually for all occurrences of the sought ref.

This is moderately cheap index-wise and moderately expensive search-wise, but I
think it strikes a good balance. If normal usage indicates otherwise, I'll
revisit.