Revamp handling of top-level compunit def.
Previously we were repeatedly "defining" the package def for each compunit in
that package, but since the imports are uses nested in that def, this was bad.
It resulted in each compunit stepping on the uses defined in the previous
compunit.

Now we create a fake def for the top of each compunit, whose id is the package
plus the filename, and we allow the imports to accumulate uses to that def.
Then we close that def when we encounter the first class, and the classes are
top-level defs in the compunit.