Revamped MapDB to use custom IdSet.
Using a mutable HashSet meant that we were often running into issues when I
forgot to clone the set at the right time and ended up with the same set in the
write queue multiple times. MapDB is full of awesome thread-safety landmines.

Now we use IdSet which is immutable, so I can't fuck up. It's also probably a
lot lower overhead, but profiling shows only a very tiny speed increase.

I really have to stop writing my own collections classes, but pulling in FJ or
something just for this one thing (and having to stay in boxed land) seemed
undesirable.