Further dealing with top-level uses.
Now we always emit a "real" def for the package, because we need all defs in a
comp unit to be enclosed by the proper package def. But we hide all the uses
that appear outside any other defs in a compunit in a special SYNTHETIC
compunit, which we'll take care to hide in the def browser UI. Blah.