Include Java files as well.
They won't be indexed, but they may be needed to prevent scalac from crapping
itself.

In a perfect world, we'd magically index the Java files and the Scala files,
but our world is so flawed and imperfect, that I'll just be happy to scrape by
with whatever minimal value we can provide.