Add standalone test for nsc.
Annoyingly it has two secret depends which are not enumerated in its POM, so we
have to magick them up lest scalac choke on its own code. I'm going to have to
hack these in if I want to get an index of scalac code for use when developing
codex, sigh.