Slightly modified version of javac.
With Java 9, we can no longer use the internals of javac that we need to make
the JavaExtractor work for Codex. As Codex is mostly being supplanted by
Language Server support, I am just copying Java 9 javac to keep this legacy
stuff working without future road bumps.

Sure would be nice if javac was shipped as a real library (without the API
stability guarantees of the full JDK, but that's fine). Clearly I should have
used JDT in the first place, alas.